Move DateTimeSearchContainer in its own file
